+++ native_locator/src/virtualenv.rs
@@ -79,7 +79,7 @@ impl Locator for VirtualEnv {
         None
     }
 
-    fn find(&self) -> Option<LocatorResult> {
+    fn find(&mut self) -> Option<LocatorResult> {
         // There are no common global locations for virtual environments.
         // We expect the user of this class to call `is_compatible`
         None
